feat(Button): revive the Reanimated Button and fix its disabled-fade gap
The shipped Button animates color transitions via @react-spring/native,
which is JS-thread-driven and updates the native view imperatively via
setNativeProps every frame. In Eigen, this caused a real bug (DI-503):
a state change that simultaneously drives this Button's disabled color
and a sibling Reanimated animation can leave the Button's color frozen
mid-transition, because the two animation systems can desync.

palette-mobile already has an unused, unexported Reanimated rewrite of
Button (ButtonNew, from #249/#253, reverted in #259 for unrelated
reasons). It only animated the pressed<->active transition though -
disabled snapped instantly with no fade, which would have been a
regression vs. the current Button for every disabled toggle in the app.

Fold disabled into the same continuous 3-state shared value as press
(DISABLED/ACTIVE/PRESSED), so the whole transition runs on the UI
thread as one interpolation and can't desync from a concurrent
Reanimated animation elsewhere on screen. Then promote ButtonNew to
replace Button, and drop @react-spring/native since nothing else in
the repo uses it.

Known non-blocking follow-ups, not addressed here:
- Icon fill color still doesn't animate on press/disabled (pre-existing
  in both the old and new Button).
- No automated assertions on interpolated color values.
