+++ src/elements/Button/Button.stories.tsx
@@ -1,10 +1,17 @@
+import { BellStrokeIcon, LinkIcon } from "@artsy/icons/native"
 import { capitalize } from "lodash"
 import { useState } from "react"
+import { Button as RNButton } from "react-native"
 import { Button, ButtonProps } from "./Button"
+import { CTAButton } from "./CTAButton"
+import { FollowButton } from "./FollowButton"
+import { LinkButton } from "./LinkButton"
 import { DataList, List } from "../../storybook/helpers"
 import { Wrap } from "../../utils/Wrap"
 import { NoUndefined } from "../../utils/types"
+import { Box } from "../Box"
 import { Flex } from "../Flex"
+import { Spacer } from "../Spacer"
 import { LinkText } from "../Text"
 import type { StoryObj } from "@storybook/react"
 
@@ -59,13 +66,12 @@ export const Sizes = () => (
 
 export const States = () => {
   const [variant, setVariant] = useState<NoUndefined<ButtonProps["variant"]>>("fillDark")
-
   return (
     <List>
       <Flex flexDirection="row" flexWrap="wrap" px={2}>
-        {variants.map((v) => (
-          <LinkText color="orange" onPress={() => setVariant(v)} mr="2">
-            {v}
+        {variants.map((variant) => (
+          <LinkText color="orange" onPress={() => setVariant(variant)} mr="2">
+            {variant}
           </LinkText>
         ))}
       </Flex>
@@ -77,6 +83,9 @@ export const States = () => {
       </Button>
       <Button variant={variant} onPress={() => console.log(`tapped`)} loading>
         Loading
+      </Button>
+      <Button variant={variant} onPress={() => console.log(`tapped`)} testOnly_pressed>
+        Pressed
       </Button>
     </List>
   )
@@ -102,21 +111,157 @@ export const Variants = () => (
 export const VariantsLoading = () => (
   <DataList
     data={variants}
-    renderItem={({ item: variant }) => {
-      if (variant !== "outlineLight") {
-        return (
-          <Button variant={variant} loading onPress={() => console.log(`tapped ${variant}`)}>
-            {variant}
-          </Button>
-        )
-      }
-      return (
-        <Flex backgroundColor="pink" py={0.5}>
-          <Button variant={variant} loading onPress={() => console.log(`tapped ${variant}`)}>
-            {variant}
-          </Button>
-        </Flex>
-      )
-    }}
+    renderItem={({ item: variant }) => (
+      <Button variant={variant} loading onPress={() => console.log(`tapped ${variant}`)}>
+        {variant}
+      </Button>
+    )}
   />
+)
+
+export const VariantsDisabled = () => (
+  <DataList
+    data={variants}
+    renderItem={({ item: variant }) => (
+      <Wrap if={variant === "outlineLight"}>
+        <Flex backgroundColor="mono100" p={1}>
+          <Wrap.Content>
+            <Button variant={variant} disabled onPress={() => console.log(`tapped ${variant}`)}>
+              {variant}
+            </Button>
+          </Wrap.Content>
+        </Flex>
+      </Wrap>
+    )}
+  />
+)
+
+export const TheFollowButton = () => {
+  const [follow, setFollow] = useState(true)
+
+  return (
+    <List>
+      <FollowButton isFollowed={follow} onPress={() => setFollow((v) => !v)} />
+      <FollowButton followCount={4} isFollowed={follow} onPress={() => setFollow((v) => !v)} />
+      <FollowButton followCount={40} isFollowed={follow} onPress={() => setFollow((v) => !v)} />
+      <FollowButton followCount={4000} isFollowed={follow} onPress={() => setFollow((v) => !v)} />
+      <FollowButton followCount={400000} isFollowed={follow} onPress={() => setFollow((v) => !v)} />
+      <FollowButton
+        followCount={40000000}
+        isFollowed={follow}
+        onPress={() => setFollow((v) => !v)}
+      />
+      <FollowButton
+        followCount={4000000000}
+        isFollowed={follow}
+        onPress={() => setFollow((v) => !v)}
+      />
+    </List>
+  )
+}
+
+export const TheCTAButton = () => (
+  <List>
+    <CTAButton onPress={() => console.log("pressed")}>cta button</CTAButton>
+  </List>
+)
+
+export const TheLinkButton = () => (
+  <List>
+    <LinkButton onPress={() => console.log("pressed")}>LinkButton</LinkButton>
+  </List>
+)
+
+export const Miscellaneous = () => (
+  <List>
+    <Button loading disabled>
+      loading and disabled
+    </Button>
+    <Button loading disabled icon={<LinkIcon />}>
+      loading and disabled with icon
+    </Button>
+    <Button block>block</Button>
+    <Flex
+      backgroundColor="orange"
+      width={400}
+      height={80}
+      alignItems="center"
+      justifyContent="center"
+    >
+      <Button variant="fillLight" icon={<LinkIcon />}>
+        left icon
+      </Button>
+    </Flex>
+    <Button icon={<LinkIcon fill="mono0" />} iconPosition="right">
+      right icon
+    </Button>
+    <Button size="small" icon={<LinkIcon fill="mono0" />} iconPosition="right">
+      Right Icon Small
+    </Button>
+    <Button
+      variant="fillDark"
+      size="small"
+      icon={<BellStrokeIcon fill="mono0" width="16px" height="16px" />}
+    >
+      Create Alert
+    </Button>
+    <Box flexDirection="row">
+      <Box width={2} height="100%" backgroundColor="green100" />
+      <Box>
+        <Button
+          size="small"
+          icon={<LinkIcon fill="mono0" />}
+          iconPosition="right"
+          longestText="this is a very long text"
+        >
+          shortest text
+        </Button>
+        <Spacer y={1} />
+        <Button
+          size="small"
+          icon={<LinkIcon fill="mono0" />}
+          iconPosition="right"
+          longestText="this is a very long text"
+        >
+          this is a very long text
+        </Button>
+      </Box>
+      <Box width={2} height="100%" backgroundColor="green100" />
+    </Box>
+    <Button icon={<LinkIcon fill="mono0" />} block iconPosition="left-start">
+      left-start aligned icon
+    </Button>
+  </List>
+)
+
+export const Playground = () => {
+  const [loading, setLoading] = useState(false)
+  const [disabled, setDisabled] = useState(true)
+  const [block, setBlock] = useState(false)
+
+  return (
+    <>
+      <RNButton title="loading" onPress={() => setLoading((v) => !v)} />
+      <RNButton title="disabled" onPress={() => setDisabled((v) => !v)} />
+      <RNButton title="block" onPress={() => setBlock((v) => !v)} />
+      <List>
+        <Button
+          loading={loading}
+          disabled={disabled}
+          block={block}
+          longestText="loading false, disabled false, block false"
+        >
+          loading {loading ? "true" : "false"}, disabled {disabled ? "true" : "false"}, block{" "}
+          {block ? "true" : "false"}
+        </Button>
+      </List>
+    </>
+  )
+}
+
+export const SpacingAround = () => (
+  <Box>
+    <Button m={2}>wow 1</Button>
+    <Button>wow 2</Button>
+  </Box>
 )
