+++ src/elements/Button/index.ts
@@ -1,4 +1,4 @@
 export * from "./Button"
 export * from "./CTAButton"
-export * from "./LinkButton"
 export * from "./FollowButton"
+export * from "./LinkButton"
