--- src/elements/ButtonNew/index.tsx
@@ -1,4 +1,0 @@
-export * from "./Button"
-export * from "./CTAButton"
-export * from "./FollowButton"
-export * from "./LinkButton"
